Move all code files into source dir
