+++ example/__tests__/App-test.tsx
@@ -4,7 +4,7 @@
 
 import 'react-native';
 import React from 'react';
-import App from '../App';
+import App from '../source/App';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
